Allow Local Config and Arbitrary Remote Config
Modify the configuration file download process to allow a local config
file to be used, or for a config file to be downloaded from an arbitrary
remote source. To use a local configuration, place the ZIP file on the
root of the internal storage, and specify the filename as *L[filename]
in the download dialog (e.g. *Lconfig.zip). To use an arbitrary remote
configuration, specify a complete URL to the source as *U[url] in the
download dialog (e.g. *Uhttps://example.com/config.zip). Also change a
few build environment settings  (Gradle & JDK versions, target SDK,
signing) to target the implementation at PSU.